+++ packages/kevin_flutter_accounts/kevin_flutter_accounts_android/test/kevin_flutter_accounts_android_test.dart
@@ -153,12 +153,8 @@ void main() {
     );
 
     const configuration = KevinAccountSessionConfiguration(state: 'state');
-    expect(
-      () async {
-        await platform.startAccountLinking(configuration);
-      },
-      throwsA(isA<PlatformException>()),
-    );
+    final result = await platform.startAccountLinking(configuration);
+    expect(result, isInstanceOf<KevinSessionUnexpectedError>());
     expect(log, <Matcher>[
       isMethodCall(
         'startAccountLinking',