+++ packages/kevin_flutter_core/kevin_flutter_core_platform_interface/lib/src/error/kevin_error_helper.dart
@@ -11,6 +11,8 @@ class KevinErrorHelper {
         return KevinSessionResultGeneralError(
           message: error.message,
         );
+      case _Errors.unexpected:
+        return KevinSessionUnexpectedError(message: error.message);
       default:
         return null;
     }
@@ -20,4 +22,5 @@ class KevinErrorHelper {
 class _Errors {
   static const general = 'general';
   static const cancelled = 'cancelled';
+  static const unexpected = 'unexpected';
 }