+++ packages/kevin_flutter_core/kevin_flutter_core_platform_interface/lib/src/model/kevin_callback_url.dart
@@ -7,7 +7,7 @@ class KevinCallbackUrl {
     required this.ios,
   });
 
-  KevinCallbackUrl.single({required String callbackUrl})
+  const KevinCallbackUrl.single({required String callbackUrl})
       : android = callbackUrl,
         ios = callbackUrl;
 }