+++ packages/kevin_flutter_core/kevin_flutter_core_platform_interface/lib/src/model/kevin_session_result.dart
@@ -2,6 +2,7 @@ abstract class KevinSessionResult {
   const KevinSessionResult();
 }
 
+/// Session result general error.
 class KevinSessionResultGeneralError extends KevinSessionResult {
   final String? message;
 
@@ -10,4 +11,12 @@ class KevinSessionResultGeneralError extends KevinSessionResult {
   });
 }
 
+/// Session initialisation/session result parsing error.
+class KevinSessionUnexpectedError extends KevinSessionResult {
+  final String? message;
+
+  const KevinSessionUnexpectedError({required this.message});
+}
+
+/// Session cancelled by user.
 class KevinSessionResultCancelled extends KevinSessionResult {}