+++ packages/kevin_flutter_core/kevin_flutter_core_platform_interface/test/kevin_error_helper_test.dart
@@ -16,6 +16,13 @@ void main() {
     expect((result as KevinSessionResultGeneralError).message, 'message');
   });
 
+  test('Parses unexpected', () {
+    final result = KevinErrorHelper.parseError(
+      PlatformException(code: 'unexpected', message: 'message'),
+    );
+    expect((result as KevinSessionUnexpectedError).message, 'message');
+  });
+
   test('Returns null on unknown error', () {
     final result =
         KevinErrorHelper.parseError(PlatformException(code: 'unknown'));