+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments/CHANGELOG.md
@@ -1,13 +1,28 @@
+## 1.0.3
+
+* Expose KevinSessionUnexpectedError - wrapper of plugin's native part's unexpected errors (wrong
+  initialisation/result parsing errors)
+* Make KevinCallbackUrl.single const constructor
+* Upgrade:
+    - `kevin_flutter_in_app_payments_platform_interface: 1.0.2`
+    - `kevin_flutter_in_app_payments_ios: 1.0.2`
+    - `kevin_flutter_in_app_payments_android: 1.0.1`
+
 ## 1.0.2
 
 * **FIX**: iOS skipBankSelection functionality fixed
-* Upgrade kevin_flutter_in_app_payments_ios to 1.0.1
+* Upgrade:
+    - `kevin_flutter_in_app_payments_ios: 1.0.1`
 
 ## 1.0.1
 
 * **BREAKING**: use KevinCallbackUrl in KevinPaymentsConfiguration instead of string
-* Add kevin_flutter_in_app_payments_android and kevin_flutter_in_app_payments_ios 1.0.0 dependencies
-* Upgrade kevin_flutter_core to 1.0.3, kevin_flutter_in_app_payments_platform_interface to 1.0.1
+* Add:
+    - `kevin_flutter_in_app_payments_android: 1.0.0`
+    - `kevin_flutter_in_app_payments_ios: 1.0.0`
+* Upgrade:
+    - `kevin_flutter_core: 1.0.3`
+    - `kevin_flutter_in_app_payments_platform_interface: 1.0.1`
 
 ## 1.0.0